+++ lizard_history/configchecker.py
@@ -8,8 +8,8 @@ logger = logging.getLogger(__name__)
 @configchecker.register
 def checker():  # Pragma: nocover
     """Verify lizard_history's demands on settings.py."""
-    if not ('tls.TLSRequestMiddleware' in
+    if not ('lizard_history.middleware.HistoryMiddleware' in
         settings.MIDDLEWARE_CLASSES):
-        logger.warn("you must add 'tls.TLSRequestMiddleware' to the list of "
-                    "MIDDLEWARE_CLASSES for the lizard_history djangoapp "
-                    "to work.")
+        logger.warn("you must add 'lizard_history.middleware.HistoryMiddleware' "
+                    "to the list of MIDDLEWARE_CLASSES for the lizard_history "
+                    "djangoapp to work.")