+++ lizard_history/models.py
@@ -1,40 +1,25 @@
 # (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.txt.
 
-from mongoengine.signals import pre_save as mongo_pre_save
-from mongoengine.signals import post_save as mongo_post_save
-from mongoengine.signals import post_delete as mongo_post_delete
-
-from django.db.models.signals import pre_save as django_pre_save
-from django.db.models.signals import post_save as django_post_save
-from django.db.models.signals import post_delete as django_post_delete
-
-from django.db.models.signals import m2m_changed
-from django.core.signals import request_finished
-from django.core.signals import got_request_exception
-
 from django.db import models
-from django.db.utils import DatabaseError
-
 from django.dispatch import receiver
-from django.contrib.sessions.models import Session
-
-from django.contrib.admin.models import LogEntry
-from django.contrib.contenttypes.models import ContentType
-from django.contrib.auth.models import Permission
-from django.contrib.auth.models import User
-from django.contrib.sites.models import Site
-from south.models import MigrationHistory
 
 from django.utils.translation import ugettext_lazy as _
 
-from lizard_history.handlers import pre_save_handler
-from lizard_history.handlers import post_save_handler
-from lizard_history.handlers import post_delete_handler
-from lizard_history.handlers import m2m_changed_handler
-from lizard_history.handlers import request_ended_handler
+from lizard_history import (
+    handlers,
+    signals,
+)
 
 import lizard_history.configchecker
 lizard_history.configchecker  # Pyflakes...
+
+# Need these models to check for excluded models
+from south.models import MigrationHistory
+from django.contrib.contenttypes.models import ContentType
+from django.contrib.auth.models import Permission
+from django.contrib.admin.models import LogEntry
+from django.contrib.sites.models import Site
+from django.contrib.auth.models import User
 
 # Some models are excluded here since they are involved in the process
 # of setting up the site
@@ -47,6 +32,7 @@ EXCLUDED_MODELS = [
     User,
 ]
 
+signals.ops_done.connect(handlers.request_handler)
 
 def _is_monitored(sender):
     """
@@ -54,107 +40,46 @@ def _is_monitored(sender):
     """
     if sender in EXCLUDED_MODELS:
         return False
-    
+    print sender 
     return MonitoredModel.objects.filter(
         app_label=sender.__module__.split('.')[0],
-        model=sender.__name__.lower()).exists()
+        model=sender.__name__.lower(),
+    ).exists()
 
 
-@receiver(django_pre_save)
-def django_pre_save_handler(sender, instance, raw, **kwargs):
-    """
-    Handle django_pre_save signal.
-    """
-    # Do nothing when loading fixtures, logging or not monitored
-    if raw or not _is_monitored(sender):
-        return
-
-    pre_save_handler(sender, instance)
+@receiver(models.signals.pre_save)
+def pre_save_handler(sender, instance, **kwargs):
+    if _is_monitored(sender):
+        kwargs.update(name='pre_save')
+        handlers.db_handler(sender, instance, **kwargs)
 
 
-@receiver(django_post_save)
-def django_post_save_handler(sender, instance, raw, **kwargs):
-    """
-    Handle django_post_save signal.
-    """
-    # Do nothing when loading fixtures, logging or not monitored
-    if raw or not _is_monitored(sender):
-        return
-
-    post_save_handler(sender, instance)
+@receiver(models.signals.post_save)
+def post_save_handler(sender, instance, **kwargs):
+    if _is_monitored(sender):
+        kwargs.update(name='post_save')
+        handlers.db_handler(sender, instance, **kwargs)
 
 
-@receiver(django_post_delete)
-def django_post_delete_handler(sender, instance, **kwargs):
-    """
-    Handle django_post_delete signal.
-    """
-    if not _is_monitored(sender):
-        return
-
-    post_delete_handler(sender, instance)
+@receiver(models.signals.pre_delete)
+def pre_delete_handler(sender, instance, **kwargs):
+    if _is_monitored(sender):
+        kwargs.update(name='pre_delete')
+        handlers.db_handler(sender, instance, **kwargs)
 
 
-@receiver(mongo_pre_save)
-def mongo_pre_save_handler(sender, document, **kwargs):
-    """
-    Handle django_pre_save signal.
-    """
-    if not _is_monitored(sender):
-        return
-
-    pre_save_handler(sender, document)
+@receiver(models.signals.post_delete)
+def post_delete_handler(sender, instance, **kwargs):
+    if _is_monitored(sender):
+        kwargs.update(name='post_delete')
+        handlers.db_handler(sender, instance, **kwargs)
 
 
-@receiver(mongo_post_save)
-def mongo_post_save_handler(sender, document, created):
-    """
-    Log a change or addition of a mongoengine document in the logentry.
-    """
-    if not _is_monitored(sender):
-        return
-
-    post_save_handler(sender, document)
-
-
-@receiver(mongo_post_delete)
-def mongo_post_delete_handler(sender, document, *args, **kwargs):
-    """
-    Handle mongo_post_delete signal.
-    """
-    if not _is_monitored(sender):
-        return
-
-    post_delete_handler(sender, document)
-
-
-@receiver(m2m_changed)
-def m2m_changed_receiver(sender, instance, action,
-                         reverse, model, pk_set, **kwargs):
-    """
-    Handle change_m2m signal.
-    """
-    if not _is_monitored(instance.__class__):
-        return
-
-    m2m_changed_handler(sender, instance, action,
-                        reverse, model, pk_set, **kwargs)
-
-
-@receiver(request_finished)
-def request_finished_receiver(sender, **kwargs):
-    """
-    Handle request_finished signal.
-    """
-    request_ended_handler()
-
-
-@receiver(got_request_exception)
-def got_request_exception_receiver(sender, **kwargs):
-    """
-    Handle request_finished signal.
-    """
-    request_ended_handler()
+@receiver(models.signals.m2m_changed)
+def m2m_changed_handler(sender, instance, **kwargs):
+    if _is_monitored(sender):
+        kwargs.update(name='m2m_changed')
+        handlers.db_handler(sender, instance, **kwargs)
 
 
 class MonitoredModel(models.Model):
@@ -179,4 +104,3 @@ class MonitoredModel(models.Model):
 
     def __unicode__(self):
         return self.name
-