+++ lizard_history/signals.py
@@ -1,0 +1,5 @@
+# (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.txt.
+
+from django.dispatch import Signal
+
+ops_done = Signal()